[#159542039] restructure folders
