+++ src/Components/NavBar/NavBar.css
@@ -1,10 +1,17 @@
 .navbar {
-    min-height: 60px;
+    min-height: 65px;
     font-size: 25px;
-    color: black;
+    color: white;
     margin-bottom: 200px;
+    background-color: black;
 }
 .nav-link{
-    color: black;
-    margin-left: 5px;
+    color: white;
+    margin-left: 6px;
 }
+.nav-link:hover {
+    opacity: 1;
+    border-bottom: 2px solid #fff;
+    text-decoration: none; 
+    margin-top: -2px;
+}
