+++ src/Components/ReturnBooks/ReturnBooks.js
@@ -1,30 +1,24 @@
 import React from 'react';
-import queryString from 'querystring';
 import axios from 'axios';
 import swal from 'sweetalert';
-import { baseURL } from '../../../helpers/baseURL';
+import baseURL from '../../helpers/baseURL';
 import '../ViewBooks/ViewBooks.css';
-import { Logout } from '../../../helpers/authUrls';
+import { Logout } from '../../helpers/authUrls';
+import { returnBook } from '../../helpers/borrowUrls';
 
-class BorrowingHistory extends React.Component {
+class ReturnBooks extends React.Component {
     constructor(props) {
         super(props);
-        const params = queryString.parse(this.props.location.search.substr(1));
         this.state = {
             books: [],
-            page: parseInt(params.page, 10),
         };
-        if (Number.isInteger(this.state.page) === false) {
-            this.state = { page: 1 };
-        }
         this.requestBooks();
-        this.paginator();
     }
 
     requestBooks() {
         const token = localStorage.getItem('token');
         axios({
-            url: `${baseURL}/users/books?page=${this.state.page}`,
+            url: `${baseURL}/users/books?returned=false`,
             method: 'get',
             headers: {
                 accept: 'application/json',
@@ -40,10 +34,9 @@ class BorrowingHistory extends React.Component {
             })
             .catch((error) => {
                 if (error.response.status === 404) {
-                    swal('No more pages.', '', 'fail')
+                    swal('No book under this Borrow Id.', '', 'fail')
                         .then(() => {
-                            const prevPage = this.state.page - 1;
-                            window.location.replace(`/hellobooks/my-history?page=${prevPage}`);
+                            window.location.replace('/hellobooks');
                         });
                 } else {
                     swal('An error occured. Please try log in again.', '', 'fail')
@@ -56,7 +49,7 @@ class BorrowingHistory extends React.Component {
 
     mapBooks = () => {
         if (this.state.books === undefined || this.state.books < 1) {
-            const BookDetails = (<h3 className='no-content'>You have not borrowed a book yet.</h3>);
+            const BookDetails = (<h3 className='no-content'>No books to return.</h3>);
             this.setState({ bookDetails: BookDetails });
         } else {
             const BookDetails = this.state.books.map(book => (
@@ -92,6 +85,12 @@ class BorrowingHistory extends React.Component {
                                         <th>Date Borrowed:</th>
                                         <td>{book.borrow_date}</td>
                                     </tr>
+                                    <button
+                                        type="button"
+                                        className="btn btn-success"
+                                        onClick={() => { this.returnBookFunc(book.borrow_id); }}>
+                                        Return This Book
+                                    </button>
                                 </tbody>
                             </table>
                         </div>
@@ -102,43 +101,19 @@ class BorrowingHistory extends React.Component {
         }
     }
 
-    paginator = () => {
-        const prevPage = this.state.page - 1;
-        const nextPage = this.state.page + 1;
-        if (this.state.page !== 1) {
-            const Paginator = (
-                <ul className="pagination">
-                    <li className="page-item"><a className="page-link" href={`/hellobooks/my-history?page=${prevPage}`}>
-                        Previous
-                    </a></li>
-                    <li className="page-item"><a className="page-link active">Page {this.state.page}</a></li>
-                    <li className="page-item"><a className="page-link" href={`/hellobooks/my-history?page=${nextPage}`}>Next</a></li>
-                </ul>
-            );
-            this.state = { paginator: Paginator };
-        } else {
-            const Paginator = (
-                <ul className="pagination">
-                    <li className="page-item"><a className="page-link disable">Page {this.state.page}</a></li>
-                    <li className="page-item"><a className="page-link" href={`/hellobooks/my-history?page=${nextPage}`}>Next</a></li>
-                </ul>
-            );
-            this.state = { paginator: Paginator };
-        }
+    returnBookFunc = (bookId) => {
+        returnBook(bookId, this.props);
     }
 
     render() {
         return (
             <div className="col-md-offset-3 col-md-6 view-books">
-                <h3 className='heading'>Your Borrowing History</h3>
+                <h3 className='heading'>Books Due</h3>
                 <div className="panel-group" id="accordion" role="tablist" aria-multiselectable="true">
                     { this.state.bookDetails }
                 </div>
-                <nav>
-                    { this.state.paginator }
-                </nav>
             </div>
         );
     }
 }
-export default BorrowingHistory;
+export default ReturnBooks;
