--- src/Components/UserBookFunctions/ViewBooks/ViewBooks.css
@@ -1,72 +1,0 @@
-.view-books {
-    margin: 60px auto;
-}
-.panel {
-    border-style: dotted;
-    background-color: #fff;
-}
-a:hover,a:focus{
-    text-decoration: none;
-    outline: none;
-   }
-   
-.panel{
-    border: none;
-    box-shadow: none;
-    border-radius: 0;
-    margin-bottom: 5px;
-   }
-   
-.panel-heading{
-    padding: 0;
-    border-radius: 0;
-    text-align: center;
-    border: none;
-   }
-   
-.panel-title a{
-    display: block;
-    padding: 15px;
-    font-size: 16px;
-    font-weight: bold;
-    color: #000;
-    background: #fff;
-    transition: all 0.5s ease 0s;
-   }
-
-.panel-body{
-    font-size: 18px;
-    color: #fff;
-    background: linear-gradient(to right, #29b0a0, #06aed5);
-    padding: 15px 20px;
-    line-height: 25px;
-    border: none;
-    position: relative;
-   }
-   
-.panel-body:after{
-    content: "";
-    position: absolute;
-    top: -30px;
-    left: 40px;
-    border: 15px solid transparent;
-    border-bottom: 15px solid #29b0a0;
-   }
-.no-content{
-    color: #fff;
-    text-align: center;
-    font-size: 40px;
-    padding-top: 20px;
-}
-table#table1{
-    width: 100%;
-    margin-bottom: 14px;
-}
-.btn{
-    color: white;
-}
-.pagination{
-    display: -ms-grid;
-    justify-content: center;
-    margin: 0 auto;
-}
