+++ src/helpers/baseURL.js
@@ -1,1 +1,2 @@
-export const baseURL = 'http://0.0.0.0:5000/api/v1';
+const baseURL = 'http://0.0.0.0:5000/api/v1';
+export default baseURL;
