+++ src/index.css
@@ -2,10 +2,6 @@ body {
   margin: 0;
   padding: 0;
   font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-  background: linear-gradient(to right, #06aed5, rgb(13, 206, 164), #01baef, #5dfdcb);
+  background: #efefef;
   color: #030301;
-}
-
-.btn-primary {
-  background-color: #00D9C0;
 }
